Keep ComboboxSelect centering internal
