+++ app/src/sandbox/combobox-center-into-view/test-browser.ts
@@ -3,9 +3,7 @@ import { withFramework } from "#app/test-utils/preview.ts";
 
 function getCenterOffset(item: Locator) {
   return item.evaluate((element) => {
-    const listbox = element.closest<HTMLElement>(
-      "[role=listbox], [data-scroll-boundary]",
-    );
+    const listbox = element.closest<HTMLElement>("[role=listbox]");
     if (!listbox) return Infinity;
     const listboxRect = listbox.getBoundingClientRect();
     const itemRect = element.getBoundingClientRect();
@@ -127,69 +125,48 @@ withFramework(import.meta.dirname, async ({ test }) => {
   });
 
   // https://github.com/ariakit/ariakit/issues/6838
-  test("preserves ancestor scrolling across a shadow slot", async ({
+  test("preserves ancestor scrolling across a same-origin iframe", async ({
     page,
-    q,
   }) => {
-    const panel = page.getByTestId("combobox-select-shadow-panel");
-    const shadowScroller = page.getByTestId("combobox-select-shadow-scroller");
-    const target = page.getByTestId("shadow-scroll-target");
+    const panel = page.getByTestId("combobox-select-iframe-panel");
+    const frame = page.frameLocator("iframe[title='Combobox select frame']");
+    const select = frame.getByRole("combobox", {
+      name: "Iframe fruit",
+      exact: true,
+    });
+    const mango = frame.getByRole("option", { name: "Mango", exact: true });
     await panel.evaluate((element) => {
       element.scrollTop = 313;
-    });
-    await shadowScroller.evaluate((element) => {
-      element.scrollTop = 180;
     });
     await test.expect
       .poll(() => panel.evaluate((element) => element.scrollTop))
       .toBe(313);
-    await test.expect
-      .poll(() => shadowScroller.evaluate((element) => element.scrollTop))
-      .toBe(180);
-    await q.button("Center shadow item").click();
 
+    await select.click();
+
+    await test.expect(select).toHaveAttribute("aria-expanded", "true");
     await test.expect
-      .poll(async () => Math.abs(await getCenterOffset(target)))
+      .poll(async () => Math.abs(await getCenterOffset(mango)))
       .toBeLessThanOrEqual(1);
-    await test.expect
-      .poll(() => shadowScroller.evaluate((element) => element.scrollTop))
-      .toBe(180);
     await test.expect
       .poll(() => panel.evaluate((element) => element.scrollTop))
       .toBe(313);
   });
 
   // https://github.com/ariakit/ariakit/issues/6838
-  test("preserves ancestor scrolling across a same-origin iframe", async ({
+  test("centers a selected item distributed through a shadow slot", async ({
     page,
     q,
   }) => {
-    const panel = page.getByTestId("combobox-select-iframe-panel");
-    const frame = page.frameLocator("iframe[title='Combobox select frame']");
-    const target = frame.getByTestId("iframe-scroll-target");
-    await panel.evaluate((element) => {
-      element.scrollTop = 313;
-    });
-    await test.expect
-      .poll(() => panel.evaluate((element) => element.scrollTop))
-      .toBe(313);
-
-    await q.button("Center iframe item").click();
-
-    await test.expect
-      .poll(async () => Math.abs(await getCenterOffset(target)))
-      .toBeLessThanOrEqual(1);
-    await test.expect
-      .poll(() => panel.evaluate((element) => element.scrollTop))
-      .toBe(313);
-  });
-
-  test("centers a selected item distributed through a shadow slot", async ({
-    q,
-  }) => {
+    const panel = page.getByTestId("combobox-select-shadow-panel");
     const select = q.combobox("Slotted fruit");
     const mango = q.option("Mango");
+    await test.expect
+      .poll(() => panel.evaluate((element) => element.scrollTop))
+      .toBe(313);
+
     await select.click();
+
     await test.expect(select).toHaveAttribute("aria-expanded", "true");
     await test.expect
       .poll(() =>
@@ -204,20 +181,10 @@ withFramework(import.meta.dirname, async ({ test }) => {
     await test.expect
       .poll(async () => Math.abs(await getAssignedCenterOffset(mango)))
       .toBeLessThanOrEqual(1);
+    await test.expect
+      .poll(() => panel.evaluate((element) => element.scrollTop))
+      .toBe(313);
     await test.expect(q.combobox("Search Slotted fruit")).toBeFocused();
-  });
-
-  test("ignores an unrelated scroll boundary", async ({ page, q }) => {
-    const scroller = page.getByTestId("unrelated-boundary-scroller");
-    await test.expect
-      .poll(() => scroller.evaluate((element) => element.scrollTop))
-      .toBe(0);
-
-    await q.button("Center with unrelated boundary").click();
-
-    await test.expect
-      .poll(() => scroller.evaluate((element) => element.scrollTop))
-      .toBe(0);
   });
 
   // https://github.com/ariakit/ariakit/issues/6838
